requirements.txt
Django==5.2
requests==2.31.0
python-dotenv==1.0.1
librosa==0.10.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1 
psycopg2-binary
cloudinary
httpx
asyncio
aiofiles
fpdf
pillow==10.2.0
